Commerce: Submit listing form

+++ commerce/auctions/models.py
@@ -15,7 +15,7 @@ class Listing(models.Model):
   description = models.TextField()
   category = models.CharField(max_length=10)
   image_url = models.URLField(blank=True)
-  cents = models.DecimalField(max_digits=9, decimal_places=2)
+  cents = models.IntegerField()
   currency = models.CharField(max_length=3, choices=CURRENCIES)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
